Move main office actions into their own controller
Moving these actions into their own controller will make the migration to
editionable worldwide organisations easier, it also follows the best practice
of not creating controller actions outside of the default 7
http://jeromedalbert.com/how-dhh-organizes-his-rails-controllers/.
